also fixing tests after reorganization
